requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
supabase
python-dotenv
httpx